requirements.txt
ansible==4.4.0
ansible-core==2.11.3
appdirs==1.4.4
asgiref==3.4.1
cffi==1.14.6
cryptography==3.4.7
distlib==0.3.2
Django==3.2.8
django-taggit==1.5.1
filelock==3.0.12
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
packaging==21.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.20
pyparsing==2.4.7
pytz==2021.3
PyYAML==5.4.1
resolvelib==0.5.4
six==1.16.0
sqlparse==0.4.2
virtualenv==20.4.7
